Refresh checkboxradio when adding/removing
Previously we only did the refresh if  prop("checked") did not
match the desired state
(e.g. if it was updated in code instead of via a click event)
We now always refresh the checkboxradio as there is a race condition
where if the user touches right between two friends it will set prop("checked")
to true on both, but only visually update the appearance of one of them.